Try remove circular ref
